docker/custom_implements/external_knowledge_api/requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
langchain-google-genai
langchain-postgres
psycopg[binary]
python-dotenv